Move cloudpathlib stuff to another module
